Backend/requirements.txt
Flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
google-generativeai==0.3.2
eventlet==0.33.3